6-puts2.c

+++ 0x05-pointers_arrays_strings/6-puts2.c
@@ -9,9 +9,9 @@
 
 void puts2(char *str)
 {
-	long long int i, z;
+	long int i, z;
 
-	long long int len = 0;
+	long int len = 0;
 
 	char *y = str;
